refactor
1. rm Global.eval_mode
2. add Global.task
3. mv type_name to ppcls.utils
4. build dataloader, loss, metric by mode
