+++ ppcls/metric/__init__.py
@@ -65,22 +65,19 @@ class CombinedMetrics(AvgMetrics):
                 metric.reset()
 
 
-def build_metrics(engine):
-    config, mode = engine.config, engine.mode
+def build_metrics(config, mode):
     if mode == 'train' and "Metric" in config and "Train" in config[
             "Metric"] and config["Metric"]["Train"]:
         metric_config = config["Metric"]["Train"]
-        if hasattr(engine.dataloader_dict["Train"],
-                   "collate_fn") and engine.dataloader_dict[
-                       "Train"].collate_fn is not None:
+        if config["DataLoader"]["Train"]["dataset"].get("batch_transform_ops",
+                                                        None):
             for m_idx, m in enumerate(metric_config):
                 if "TopkAcc" in m:
                     msg = f"Unable to calculate accuracy when using \"batch_transform_ops\". The metric \"{m}\" has been removed."
                     logger.warning(msg)
                     metric_config.pop(m_idx)
         train_metric_func = CombinedMetrics(copy.deepcopy(metric_config))
-    else:
-        train_metric_func = None
+        return train_metric_func
 
     if mode == "eval" or (mode == "train" and
                           config["Global"]["eval_during_train"]):
@@ -97,7 +94,4 @@ def build_metrics(engine):
             else:
                 metric_config = [{"name": "Recallk", "topk": (1, 5)}]
             eval_metric_func = CombinedMetrics(copy.deepcopy(metric_config))
-    else:
-        eval_metric_func = None
-
-    return train_metric_func, eval_metric_func
+        return eval_metric_func
